Move to C++.
